Fix error view not getting focus on signup screen
Commit cf41f4c768a32cdb01d71617c2713ab79ea55f08 added a popup dialog in
addition to the field errors. However, it caused a regression in that it
failed to actually focus on the error field, and instead focused on the
first visible field instead. This commit fixes the regression by
readding the logic to focus on the first invalid field entry.

An additional redundant call to request the focus on the concerned view
that has been added in that commit has also been removed, as well as
some minor reordering of the logic to be more consistent.